requirements.txt
boto3
psycopg2-binary
uvicorn
fastapi
python-dotenv
python-multipart
